video annotation lots of blinking
this person had lots of blinking and moving around (I think it was a
really young participant) but then there are momemts when both head and
eyes are very still/stable. could be a good proxy, since the eyes will
be all over the place and eye tracking may not work.