reverted supersampled_worldCam_rawLiveVid
the world cam was catching the last few frames of the "Please Center Your Eye" phase because of the display lag (due to hardware) between when Bonsai sends the signal to show "Do not move head" phase (and simultaneously telling the world cam to start recording) and when the signal from Bonsai is actually resolved by hardware setup and the display monitor shows the first frame of "Do not move head" phase.

This lag will be hard coded into psa02_PupilSizeVLum_scatter.py, just before downsampling the world cam and raw live stim arrays.